# Conflicts:
#	include/PgSQL_Monitor.hpp
#	lib/PgSQL_HostGroups_Manager.cpp
#	lib/PgSQL_Monitor.cpp
#	lib/ProxySQL_Admin.cpp
#	lib/ProxySQL_Config.cpp

+++ test/tap/tests/firewall_commands1-t.cpp
@@ -78,7 +78,7 @@ int main() {
 
 
 
-	plan(queries.size());
+	plan(static_cast<int>(queries.size()));
 
 	for (std::vector<std::string>::iterator it2 = queries.begin(); it2 != queries.end(); it2++) {
 		MYSQL* proxysql_admin = mysql_init(NULL); // local scope . We intentionally create new connections
